[Issue #2] Temporaray files are now deleted
Issue was caused due to gnuplot creating plots only on destruction of
object of which a gnuplot object was a member. Since, gnuplot operation
is called only on object destruction, it is better to create a temprary
gnuplot object and let it go out of scope at the end of the function.